+++ webpack.config.js
@@ -1,4 +1,9 @@
+const fs = require("fs");
 const path = require("path");
+const webpack = require("webpack");
+
+/* Read license */
+const license = fs.readFileSync("./LICENSE", "utf-8");
 
 module.exports = {
 	entry: {
@@ -8,4 +13,5 @@ module.exports = {
 		filename: "all-sites-menu.js",
 		path: path.resolve(__dirname, "include"),
 	},
+	plugins: [new webpack.BannerPlugin(license)],
 };